requirements.txt
mysql-connector-python
dotenv
pandas
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-api-python-client
numpy 
